chore: fixing services v2
