+++ proj_backend/src/validations/userRole.js
@@ -1,0 +1,11 @@
+const Joi = require('joi');
+
+const validate = (request) => {
+  const schema = {
+    userId: Joi.required(),
+    roleId: Joi.required()
+  }
+  return Joi.valid(request, schema);
+};
+
+exports = validate;
