--- proj_backend/src/validations/userRoles.js
@@ -1,10 +1,0 @@
-const Joi = require('joi');
-
-const validate = (request) => {
-    const schema = {
-        userId: Joi.required(),
-        roleId: Joi.required()
-    }
-}
-
-exports = validate;
